+++ src/adapters/qq/pool.ts
@@ -15,7 +15,6 @@ interface QQAdapterPoolOptions {
   redisUrl: string;
   registryPrefix: string;
   refreshIntervalSec?: number;
-  registryTtlSec?: number;
   logger?: Logger;
 }
 
@@ -42,7 +41,6 @@ export class QQAdapterPool implements PlatformAdapter {
       redisUrl: options.redisUrl,
       prefix: options.registryPrefix,
       refreshIntervalSec: options.refreshIntervalSec,
-      ttlSec: options.registryTtlSec,
       logger: this.logger,
     });
   }